backend/requirements.txt
Flask
google-generativeai
python-dotenv
PyPDF2
python-docx
Pillow
bleach
flask-cors
gunicorn
rank-bm25